update_nov_1

+++ tool/bin/Communities/run_Communities.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
+
 # STEP 0
-echo '  0/2 - Preparing environment'
-for file in input/layers/*; do mv "$file" "${file/.*/.gr}"; done
-mkdir -p bin/Communities/src/genes; mkdir -p bin/Communities/src/networks; cp input/layers/*.gr bin/Communities/src/networks; cp output/metaDEGs/metaDEGs/degs_names.txt bin/Communities/src/genes/input_genes.txt
+echo '      - Preparing environment'
+for file in input/layers/*; do sed -i 's/,/ /g' $file ; mv "$file" "${file/.*/.gr}"; done
+mkdir -p bin/Communities/src/genes; mkdir -p bin/Communities/src/networks; cp input/layers/*.gr bin/Communities/src/networks; cp output/metaDEGs/degs_names.txt bin/Communities/src/genes/input_genes.txt
 # arguments to variables
 Molti_modularity=$1 
 Molti_Louvain=$2
@@ -10,12 +11,24 @@ pushd bin/Communities/ >& /dev/null
 rm -r -f output/*; mkdir -p output/tmp
 
 # STEP 1
-echo '  1/2 - Describing communities from multilayer networks'
+echo '      - Defining Communities'
 networks=$(ls src/networks/*.gr)
-src/MolTi-DREAM-master/src/molti-console -p ${Molti_modularity} -r ${Molti_Louvain} -o output/tmp/communities ${networks} >& /dev/null
+
+if [ $Molti_modularity = 1 ] && [ $Molti_Louvain = 0 ]; then
+   ./src/MolTi-DREAM/src/molti-console -o output/tmp/communities ${networks} >& /dev/null
+else
+  ./src/MolTi-DREAM/src/molti-console -p ${Molti_modularity} -r ${Molti_Louvain} -o output/tmp/communities ${networks} >& /dev/null
+fi
+
+
+# Check if molti runned correctly
+if [[ ! -f output/tmp/communities ]] ; then
+    echo 'Problem found when running MolTi, please, check documentation.'
+    exit 1
+fi
 
 # STEP 2
-echo '  2/2 - Finding genes in COMM'
+echo '      - Analysing gene distribution in communities'
 ## check genes input name
 cp src/genes/input_genes.txt output/tmp/degs.txt; sed -i 's/$/;/g' output/tmp/degs.txt
 ## Obtain matches
@@ -31,7 +44,9 @@ for clusterid in $(ls output/tmp/cluster_*.txt | cut -d"_" -f2 | cut -d"." -f1);
     done <output/tmp/cluster_${clusterid}.txt
 done
 
+sed -ri 's/.* ClusterID: (.*)/\1/' output/tmp/communities_effectif.csv
+
 ## Clean result
-mv output/tmp/degs.txt output/degs_communities.txt; mv output/tmp/communities output/communities.txt; rm -r -f output/tmp
+mv output/tmp/degs.txt output/degs_communities.txt; mv output/tmp/communities output/molti_output.txt; mv output/tmp/communities_effectif.csv output/size_communities.txt ;mkdir -p output/clusters; mv output/tmp/cluster_*.txt output/clusters/ ; rm -r -f output/tmp
 popd >& /dev/null
-mkdir -p output/Communities; mv bin/Communities/output/* output/Communities; rm -r -f bin/Communities/src/genes; rm -r -f bin/Communities/src/networks; rm -r -f bin/Communities/output
+mkdir -p output/Communities; mv bin/Communities/output/* output/Communities
